bug free.

+++ Database/src/main/java/engine/transaction/function/AVG.java
@@ -2,7 +2,7 @@ package engine.transaction.function;
 
 public class AVG extends Function {
 
-    public AVG(Double delta) {
+    public AVG(Integer delta) {
 
         this.delta_double = delta;
     }
